image-to-pdf-converter/requirements.txt
Pillow==10.0.1
reportlab==4.0.7 